+++ tests/test.py
@@ -1,10 +1,10 @@
 import json
 from pathlib import Path
 from unittest.mock import patch
-
 from io import StringIO
 import unittest
-from workbench_lite import WorkbenchLite
+
+import workbench_core
 
 
 def get_snapshot(file_name: str) -> dict:
@@ -14,25 +14,25 @@ def get_snapshot(file_name: str) -> dict:
 
 
 class TestWorkbenchLite(unittest.TestCase):
-    workbench_lite = WorkbenchLite()
+    workbench = workbench_core.WorkbenchCore()
 
     @patch('sys.stdout', new_callable=StringIO)
     def test_generate_snapshot__check_stdout(self, stdout):
         """
         Test to check if function generate snapshot works properly.
         """
-        self.workbench_lite.generate_snapshot()
+        self.workbench.generate_snapshot()
         expected_stdout = 'Snapshot json saved!\n'
         self.assertEqual(stdout.getvalue(), expected_stdout)
 
     def test_post_snapshot__response_code_201(self):
         snapshot = get_snapshot('snapshot.full')
-        r = self.workbench_lite.post_snapshot(snapshot)
+        r = self.workbench.post_snapshot(snapshot)
         assert r['code'] == 201
 
     def test_post_snapshot__response_code_422(self):
         snapshot = get_snapshot('snapshot.empty')
-        r = self.workbench_lite.post_snapshot(snapshot)
+        r = self.workbench.post_snapshot(snapshot)
         assert r['code'] == 422